+++ src/components/manifold-copy-credentials/manifold-copy-credentials-happo.ts
@@ -1,0 +1,4 @@
+export const loading = async () => {
+  const copy = document.createElement('manifold-copy-credentials');
+  document.body.appendChild(copy);
+};