+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - Enabled custom loading indicators.
+- Added `<manifold-copy-credentials>` for quickly-copying secrets to a user’s clipboard
 
 ### Fixed